requirements.txt
django>=4.2
djangorestframework>=3.14
django-filter>=23.0
djangorestframework-simplejwt>=5.3
